+++ src/unauthorized/index.js
@@ -1,6 +1,7 @@
 import { assign } from '@ctx-core/object'
 import { HttpError } from '../http_error/index.js'
 /** @typedef {import('../_types').argument__error_o_T}argument__error_o_T */
+export const unauthorized__default__message = 'Unauthorized'
 /**
  * @param {string}[message]
  * @param {argument__error_o_T}[error_o]
@@ -11,7 +12,9 @@ export function unauthorized_error_(
 	message,
 	error_o
 ) {
-	return assign(new UnauthorizedError(message), error_o)
+	return assign(
+		new UnauthorizedError(message ?? unauthorized__default__message),
+		error_o)
 }
 /**
  * Throws an unauthorized error (HTTP 401)
@@ -34,7 +37,7 @@ export class UnauthorizedError extends HttpError {
 	constructor(message) {
 		super(message)
 		this.name = 'UnauthorizedError'
-		this.http__message = 'Unauthorized'
+		this.http__message = unauthorized__default__message
 		if (!this.message) this.message = this.http__message
 	}
 }